update ban hang

+++ TheHans/src/app/dto/HoaDonDTO.java
@@ -18,9 +18,9 @@ public class HoaDonDTO {
     private String maHoaDon;
     private String tenNguoiNhan;
     private String diaChi;
-    private Long tienKhachTra;
-    private Long tienThuaLai;
-    private Long thanhTien;
+    private Double tienKhachTra;
+    private Double tienThuaLai;
+    private Double thanhTien;
     private boolean trangThaiXoa;
     private Date ngayTao;
     private Date ngaySuaCuoi;
@@ -35,7 +35,7 @@ public class HoaDonDTO {
 
     
     
-    public HoaDonDTO(int hoaDonId, int id_KhachHang, int id_NhanVien, String maHoaDon, String tenNguoiNhan, String diaChi, Long tienKhachTra, Long tienThuaLai, Long thanhTien, boolean trangThaiXoa, Date ngayTao, Date ngaySuaCuoi, String ghiChu, String tenNhanVien, String tenKhachHang, String sdtKhachHang) {
+    public HoaDonDTO(int hoaDonId, int id_KhachHang, int id_NhanVien, String maHoaDon, String tenNguoiNhan, String diaChi, Double tienKhachTra, Double tienThuaLai, Double thanhTien, boolean trangThaiXoa, Date ngayTao, Date ngaySuaCuoi, String ghiChu, String tenNhanVien, String tenKhachHang, String sdtKhachHang) {
         this.hoaDonId = hoaDonId;
         this.id_KhachHang = id_KhachHang;
         this.id_NhanVien = id_NhanVien;
@@ -54,7 +54,7 @@ public class HoaDonDTO {
         this.sdtKhachHang = sdtKhachHang;
     }
 
-    public HoaDonDTO(int hoaDonId, int id_KhachHang, int id_NhanVien, String maHoaDon, String tenNguoiNhan, String diaChi, Long tienKhachTra, Long tienThuaLai, Long thanhTien, boolean trangThaiXoa, Date ngayTao, Date ngaySuaCuoi, String ghiChu, String tenNhanVien, String tenKhachHang, String sdtKhachHang, String hinhThucThanhToan) {
+    public HoaDonDTO(int hoaDonId, int id_KhachHang, int id_NhanVien, String maHoaDon, String tenNguoiNhan, String diaChi, Double tienKhachTra, Double tienThuaLai, Double thanhTien, boolean trangThaiXoa, Date ngayTao, Date ngaySuaCuoi, String ghiChu, String tenNhanVien, String tenKhachHang, String sdtKhachHang, String hinhThucThanhToan) {
         this.hoaDonId = hoaDonId;
         this.id_KhachHang = id_KhachHang;
         this.id_NhanVien = id_NhanVien;
@@ -124,27 +124,27 @@ public class HoaDonDTO {
         this.diaChi = diaChi;
     }
 
-    public Long getTienKhachTra() {
+    public Double getTienKhachTra() {
         return tienKhachTra;
     }
 
-    public void setTienKhachTra(Long tienKhachTra) {
+    public void setTienKhachTra(Double tienKhachTra) {
         this.tienKhachTra = tienKhachTra;
     }
 
-    public Long getTienThuaLai() {
+    public Double getTienThuaLai() {
         return tienThuaLai;
     }
 
-    public void setTienThuaLai(Long tienThuaLai) {
+    public void setTienThuaLai(Double tienThuaLai) {
         this.tienThuaLai = tienThuaLai;
     }
 
-    public Long getThanhTien() {
+    public Double getThanhTien() {
         return thanhTien;
     }
 
-    public void setThanhTien(Long thanhTien) {
+    public void setThanhTien(Double thanhTien) {
         this.thanhTien = thanhTien;
     }
